requirement.txt
Flask
pyttsx3
requests
wikipedia
python-dotenv
google-generativeai
